+++ CHANGELOG.md
@@ -5,6 +5,8 @@ QuantumCircuitOpt.jl Change Log
 - Moved all the model options (including valid constraints) from `qc_model.jl` to `types.jl` into `QCModelOptions`, a struct form 
 - Streamlined default options for `QCModelOptions`
 - Moved `relax_integrality`, `time_limit`, `objective_slack_penalty` and `optimizer_log` options from `data.jl` to `QCModelOptions`
+- Removed `identify_real_gates`, `optimizer` and `optimizer_presolve` options from `params` and updated `solvers.jl`
+- Default recognition of real elementary and target gates, and implements a compact MIP
 - Updated docs and unit tests to reflect above changes
 
 ### v0.3.1